+++ lib/src/utils/web/web_utils.dart
@@ -1,0 +1,1 @@
+export 'web_utils_stub.dart' if (dart.library.html) 'web_utils_web.dart';